Files module : prevent crash if file or directory doesn't exist

+++ modules/Files.csx
@@ -5,7 +5,7 @@ public class Files
 	//Copy folder content to another folder
     public static void CopyFolder(string sourceDirName, string destDirName, bool copySubDirs = true, bool overwrite = false)
 	{
-		BuildHelper.DisplayAndLog("Copying directory '" +  sourceDirName + "' to '" +  destDirName + "'...");
+		BuildHelper.DisplayAndLog("Copying directory '" + sourceDirName + "' to '" + destDirName + "'...");
 		// Get the subdirectories for the specified directory.
         DirectoryInfo dir = new DirectoryInfo(sourceDirName);
 		DirectoryInfo[] dirs = dir.GetDirectories();
@@ -45,7 +45,7 @@ public class Files
 	{
 		if(System.IO.File.Exists(filePath))
 		{
-			BuildHelper.DisplayAndLog("Deleting file '" +  filePath + "'...");
+			BuildHelper.DisplayAndLog("Deleting file '" + filePath + "'...");
 			System.IO.File.Delete(filePath);
 		}
 	}
@@ -53,12 +53,18 @@ public class Files
 	//Get all the files of a directory following a regex patern
 	public static string[] GetFilesWithPattern(string parentDirectoryPath, string filePattern, bool subdirectories = false)
 	{
-		return System.IO.Directory.GetFiles(parentDirectoryPath, filePattern, subdirectories ? System.IO.SearchOption.AllDirectories : System.IO.SearchOption.TopDirectoryOnly );
+		if(!System.IO.Directory.Exists(parentDirectoryPath))
+			return new string[0];
+
+		return System.IO.Directory.GetFiles(parentDirectoryPath, filePattern, subdirectories ? System.IO.SearchOption.AllDirectories : System.IO.SearchOption.TopDirectoryOnly);
 	}
 
 	//Delete all the files of a directory following a regex patern
 	public static void DeleteFilesWithPattern(string parentDirectoryPath, string filePattern)
 	{
+		if(!System.IO.Directory.Exists(parentDirectoryPath))
+			return;
+
 		foreach(var directory in GetFilesWithPattern(parentDirectoryPath, filePattern))
 			DeleteFile(directory);
 	}
@@ -66,6 +72,9 @@ public class Files
 	//Delete all the subdirectories of a directory following a regex patern
 	public static void DeleteDirectoriesWithPattern(string parentDirectoryPath, string directoryPattern)
 	{
+		if(!System.IO.Directory.Exists(parentDirectoryPath))
+			return;
+
 		foreach(var directory in System.IO.Directory.GetDirectories(parentDirectoryPath, directoryPattern))
 			DeleteDirectory(directory);
 	}
@@ -75,13 +84,16 @@ public class Files
 	{
 		if(System.IO.Directory.Exists(directoryPath))
 		{
-			BuildHelper.DisplayAndLog("Deleting directory '" +  directoryPath + "'...");
+			BuildHelper.DisplayAndLog("Deleting directory '" + directoryPath + "'...");
 			System.IO.Directory.Delete(directoryPath, true);
 		}
 	}
 
 	public static void ReplaceText(string filePath, string regex, string newText)
 	{
+		if(!System.IO.File.Exists(filePath))
+			return;
+
 		BuildHelper.DisplayAndLog("Replacing text in file '" +  filePath + "'...");
 		File.WriteAllText(filePath, Regex.Replace(File.ReadAllText(filePath), regex, newText));
 	}
